+++ LuaBind/LuaBind_Main.cpp
@@ -3,7 +3,6 @@
 // author @zoloypzuo
 #include <sol/sol.hpp>
 
-void LuaBind_Config(sol::state &luaState);
 
 void LuaBind_ThirdParty(sol::state &luaState);
 
@@ -14,7 +13,6 @@ void LuaBind_Renderer(sol::state &luaState);
 void LuaBind_Sandbox(sol::state &luaState);
 
 void LuaBind_Main(sol::state &luaState) {
-    LuaBind_Config(luaState);
     LuaBind_ThirdParty(luaState);
     LuaBind_Core(luaState);
     LuaBind_Renderer(luaState);
