+++ Script/Lua/main.lua
@@ -3,55 +3,10 @@
 -- author @zoloypzuo
 print("start running main.lua")
 
--- LUA_PATH
--- config
-package.path = package.path .. ";" .. CONFIG_DIR .. "/?.lua"
-
--- resource
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Entities/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Entities/Materials/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Entities/Models/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Entities/Textures/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Fonts/?.lua"
-package.path = package.path .. ";" .. RESOURCE_DIR .. "/Shaders/?.lua"
-
--- script
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/common/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/engine/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/framework/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/framework/debug/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/prefabs/?.lua"
-package.path = package.path .. ";" .. SCRIPT_DIR .. "/Lua" .. "/scriptlibs/?.lua"
-
-local ENABLE_LUA_PANDA = true
-if ENABLE_LUA_PANDA then
-    local ZBS = "D:\\Installed\\ZeroBraneStudio"
-    package.path = package.path .. ";" .. ZBS .. "\\lualibs\\?\\?.lua;" .. ZBS .. "\\lualibs\\?.lua"
-    package.cpath = package.cpath .. ";" .. ZBS .. "\\bin\\?.dll;" .. ZBS .. "\\bin\\clibs\\?.dll"
-    require("debugger.LuaPanda").start("127.0.0.1", 8818)
--- require("mobdebug").start()
-end
-
-DUMP_LUA_PATH = false
-if DUMP_LUA_PATH then
-    print("=== Start dump path info")
-    print("package.path", package.path)
-    print("package.cpath", package.cpath)
-    print("PATH", os.getenv("PATH"))
-    print("SCRIPT_DIR", SCRIPT_DIR)
-    print("=== End dump path info")
-end
-
-global = function(name)
-end
-
 global("CWD")
 CWD = SCRIPT_DIR .. "/Lua"
 
 -- global require
---require("strict")
 
 require("imgui.imgui_consts")
