+++ Script/Lua/main_function.lua
@@ -316,18 +316,3 @@ function LoadAvatar()
     avatar.entity:AddFreeMove()
     avatar.entity:AddFreeLook()
 end
-
-global("ConfigCache")
-ConfigCache = {}
-function registerConfigClass(klass_name)
-    local klass = _G[klass_name]
-    local mt = getmetatable(klass)
-    mt.__call = function(class_tbl, data)
-        local o = class_tbl.new()
-        for k, v in pairs(data) do
-            o[k] = v
-        end
-        ConfigCache[#ConfigCache + 1] = o
-        return o
-    end
-end
